requirements.txt
Flask
flask_socketio
flask_cors
google-generativeai
ipython
python-dotenv
sentence_transformers
tika
pinecone-client
uuid
Gunicorn
